fixed chatroom size in dashboard
