+++ src/components/Chat.jsx
@@ -24,67 +24,73 @@ const useStyles = makeStyles({
   },
   chatSection: {
     width: '100%',
-    height: '83vh'
+    height: '83vh',
   },
   headBG: {
-      backgroundColor: '#e0e0e0'
+    backgroundColor: '#e0e0e0',
   },
   borderRight500: {
-      borderRight: '1px solid #e0e0e0'
+    borderRight: '1px solid #e0e0e0',
   },
   messageArea: {
-    height: '100vh',
-    overflowY: 'auto'
+    height: '72vh',
+    overflowY: 'auto',
   },
   roomBox: {
     color: '#40637E',
     fontWeight: 'bold',
-    fontSize: "25px"
+    fontSize: '25px',
   },
-  avatar : {
+  avatar: {
     color: '#fff',
-    backgroundColor: "#40637E",
-  }
+    backgroundColor: '#40637E',
+  },
 });
 
-const Chat = ({name, room}) => {
+const Chat = ({ name, room }) => {
   const classes = useStyles();
   const [messages, typeMsg, sendNewMessage, sendTypingMsg] = useSocket(
     name,
     room
   );
 
-
-
   return (
     <div>
       <Grid container component={Paper} className={classes.chatSection}>
-          <Grid item xs={3} className={classes.borderRight500}>
-              <List>
-                  <ListItem button >
-                      <Typography className={classes.roomBox}>{room}</Typography>
-                  </ListItem>
-                  <Divider />
-                  <ListItem button >
-                      <ListItemIcon>
-                          <Avatar className={classes.avatar}>{name[0]}</Avatar>
-                      </ListItemIcon>
-                      <ListItemText >{name} </ListItemText>
-                  </ListItem>
-              </List>
+        <Grid item xs={3} className={classes.borderRight500}>
+          <List>
+            <ListItem button>
+              <Typography className={classes.roomBox}>{room}</Typography>
+            </ListItem>
+            <Divider />
+            <ListItem button>
+              <ListItemIcon>
+                <Avatar className={classes.avatar}>{name[0]}</Avatar>
+              </ListItemIcon>
+              <ListItemText>{name} </ListItemText>
+            </ListItem>
+          </List>
+        </Grid>
+        <Grid container xs={9} direction="column">
+          <Grid item className={classes.messageArea}>
+            <Messages
+              // style={{ height: '25vh' }}
+              messages={messages}
+              name={name}
+              typeMsg={typeMsg}
+            />
           </Grid>
-          <Grid item xs={9} className={classes.messageArea}>
-                <Messages messages={messages} name={name} typeMsg={typeMsg} />
-              <Divider />
-     
-              <InputBox sendNewMessage={sendNewMessage} sendTypingMsg={sendTypingMsg}/>
+          <Divider />
+          <Grid item>
+            <InputBox
+              // style={{ height: '15vh' }}
+              sendNewMessage={sendNewMessage}
+              sendTypingMsg={sendTypingMsg}
+            />
           </Grid>
+        </Grid>
       </Grid>
     </div>
-);
-
-
-
-
+  );
 };
 export default Chat;
